refactor: rename dir utils to config
